requirements.txt
opencv-python >= 4.7.0.0
Pillow >= 9.5